move all `extract_array` logic into `evolve`
Having it in it's own function, especially when it did extra
things with the BH bins, made this function a bit obscure.
It sounded like all it did was "extract" the arrays, but it
actually does more.

There wasn't really a need for a separate function for it,
and it makes sense for any logic which directly affects what is
outputted to be visible in the same place.