delete a handful of unused files
removes various modules not typically used. This includes both
completely empty, auto-created files like `cli` and scripts that
seem to have been created for a specific purpose (often with local
and unusable paths) but are not for general usage.

This files will remain, if desired, in older versions of the
package, but should not be kept around here.