fix divide-by-zero bug in `Pk` helper function
`Pk` is a helper function representing the analytic solution of
the integral over the mass. This solution takes a different form
when `a == k`, as the typical solution would have a divide by
zero error.

The previous fix for this assumed a couple erronous things;
firstly that one or the other solution would be valid for all
inputs submitted to the method in a single call, and secondly that,
when called on arrays, numpy would raise a divide by zero array,
rather than just giving a nan.

This commit adds a more robust, but hopefully still efficient, fix.