fix bug in maximum possible final WD masses
In similar fashion to a77ae3c, the final maximum mass possible
from the WD IFMR was sometimes being predicted as lower than
what was actually possible, as the polynomial used to interpolate
it sometimes happens to turn down near the end of the domain.

This is completely unphysical, and solely the fault of using
polynomials to represent this IFMR, and it was predicted this
might happen.
With how the new remnant bins are being created, using `WD_mf`,
this bug finally became breaking.

The simple fix here simply determines what the actual local max
of the polynomial is and uses that