fix bugs surrounding out-of-bounds mto computation
Now that the usual check for time vs `tms_u` before adjusting the
mto bounds are removed, the actual copmutation of the mto and the
determination/adjustment of the corresponding mass bounds, needs
to be adjusted to correctly handle cases outside the norm.

The `compute_mto` equation is invalid below the maximum at `a[0]`,
and in the past relied on the fact that the lowest mass bound
was typically above that, but it was never guaranteed. Now it will
explicitly go to inf, and the bins will not be adjusted when
outside the bounds