fix WD mf upper bounds taken from polynomial
The polynomial roots method recently introduced had problems
with some polynomials. At some metallicities the function goes
off to infinity below ~1. I'm guessing the data the polynomial
was fit on ended around there. This commit simply applies a lower
limit to the possible max position (at 1).
The upper bound is also evaluated now, to check if it is the
max position, since it isn't guaranteed to be part of the roots
otherwise.