rewritepack/unpack to be more speed efficient
While using the numpy `r_` and `split` functions is neater, and
probably fast enough, rewriting these functions to more
manually manipulate the arrays is faster, by avoiding creating
a number of interim arrays behind the scenes. This reduces it by
a few hundred ms, meaning the speed is nearly the same as it was
before all the mass bin changes.