make EvolvedMF classes use new IMF to init
Changes the arguments required to initialize the EvolvedMF class
and subclasses to instead take an IMF instance directly, rather
than the break masses and slopes themselves.
This old signature is instead split into a classmethod which
recreates the old behaviour.

This allows for re-using an IMF class externally while also
initializing an EvolvedMF with it, and allows for more customizable
init to these classes, for example allowing the mass bin breaks
to differ from the IMF breaks.