fix floating-point bug in ejecting all BHs
When BHret~0 (and natal kicks are turned on) floating point errors
can sometimes cause the amount of BHs to be dynamically ejected to
be larger than the total amount available, causing an untrue
error and misleading error message.
This commit fixes the bug by adding a pre-check where the kick
algorithms are skipped over, if the result will be basically 0
BHs retained.