fix bug in BH kicks when more than one `tout` age
Because BH kicks are applied post facto, not using a copied
`y` array to setup all the mass bins was wreaking havoc with the
ODE solver, when the kicks were applied multiple times.
In the past this had been discretely hidden by the `Nmin` check
in `dyn_eject_BH`, but the BH results would still be wrong.

While how we handle BH kicks is still technically problematic, and
extra care is required when using these methods to look at BHs
over multiple times, at least this fix will allow the ODE to
converge.