add more manunal BH mf bound determinations
For the broken power law IFMR prescription, the discontinuity at
the break masses meant that numerical solvers were unable to find
the minimum/maximum values to high enough precision. Mostly because
the ODE solver likes to look at mto values a hair above the bin
edges.
This simply manually checks each power law at the bounds only for
the max and min values. I think, given the form of the power law,
that this is fine because that min or max should likely always
appear at the boundaries anyways.