update SSE IFMR grid
The grid of SSE models (based on the updates of Banerjee+2020) was
recomputed, in order to fix a few things with the old grid.
First was to nail down exactly what SSE parameters were used,
then to add support for both the delayed and rapid Fryer2012
prescriptions, then to increase the timescale resolution (pts1),
and finally to fix a bug in the naming of the grid files using the
wrong solar metallicity value (which had negligible effects until
you reached Z>~0.01).

This commit adds the new grids and support for them in `ifmr`,
and renames the method to "Banerjee20" (as the Fryer2012
prescriptions are used by many models).