add new module for sev-only prescriptions
This adds the `sev` module which contains a couple classes
designed to be used entirely independently of the normal `EvolvedMF`
to track the impacts of stellar evolution in a continuous
(non-binned) manner.
These have been tested to match the results of the similar derivatives
in `EvolvedMF`, as long as the rtol of a solver is small enough.