Fix bugs surrounding stored kick stats
This commit fixes a few things to do with the stored `_kick_stats`
in EvolvedMF when the kicks were not being computed normally.
Most notably, before there was a bug when ejecting all BHs, because
the kick stats were being written manually in EvolvedMF. Now, there
is a new default "no kicks have been applied" constructor to the
`KickStats` class, and the kicks are being computed (and then ignored)
in the case when `BH_ret_dyn` is ejecting everything, so the stats
can be stored. While this is technically less efficient, the kicks
are quite fast and these stats are still interesting when ejecting
all the rest of the BHs dynamically anyway.